+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added Chinese (Simplified) translation ([#24](https://codeberg.org/silverpill/mitra-web/pulls/24)).
 
+### Fixed
+
+- Fixed background color of search input in webkit browsers.
+
 ## [2.24.0] - 2024-06-30
 
 ### Added